PETSc Vec

Base.setindex! — Method

Base.setindex!(vec::Vec, value::Number, row::Integer)

row must be in [1,size(vec)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling VecSetValue fails.

PetscWrap.create_vector — Function

create_vector(
    comm::MPI.Comm = MPI.COMM_WORLD;
    nrows_loc = PETSC_DECIDE,
    nrows_glo = PETSC_DECIDE,
    autosetup = false,
)

Create a Vec vector of global size (nrows_glo).

PetscWrap.get_range — Method

get_range(vec::Vec)

Wrapper to VecGetOwnershipRange

However, the result (rstart, rend) is such that vec[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc.VecGetOwnershipRange result where the indexing starts at zero and where rend-1 is last row handled by the local processor.

PetscWrap.get_urange — Method

get_urange(vec::Vec)

Provide a UnitRange from the method get_range.

PetscWrap.set_local_to_global! — Method

Wrapper to ISSetLocalToGlobalMapping

1-based indexing

PetscWrap.set_value! — Function

set_value!(vec::Vec, row, value, mode::InsertMode = INSERT_VALUES)

Wrapper to VecSetValue, using julia 1-based indexing.

PetscWrap.set_value_local! — Function

set_value_local!(vec::Vec, row, value, mode::InsertMode = INSERT_VALUES)

Wrapper to VecSetValueLocal, using julia 1-based indexing.

PetscWrap.set_values! — Function

set_values!(vec::Vec, rows, values, mode::InsertMode = INSERT_VALUES)
set_values!(vec::Vec, values)

Wrapper to VecSetValues, using julia 1-based indexing.

PetscWrap.set_values_local! — Function

set_values_local!(vec::Vec, rows, values, mode::InsertMode = INSERT_VALUES)
set_values_local!(vec::Vec, values)

Wrapper to VecSetValuesLocal, using julia 1-based indexing.

PetscWrap.vec2array — Method

vec2array(vec::Vec)

Convert a Vec into a Julia Array. Allocation is involved in the process since the Vec allocated by PETSC is copied into a freshly allocated array. If you prefer not to allocate memory, use VectGetArray and VecRestoreArray

PetscWrap.vec2file — Function

vec2file(vec::Vec, filename::String, format::PetscViewerFormat = PETSC_VIEWER_ASCII_CSV, type::String = "ascii")

Write a Vec to a file.

PETSc Mat

Base.setindex! — Method

row and col must be in [1,size(mat)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling MatSetValue fails.

PetscWrap.assemble! — Function

Wrapper to `MatAssemblyBegin` and `MatAssemblyEnd` successively.

PetscWrap.create_composite_add — Method

Wrapper to MatCreateComposite using the "alternative construction" from the PETSc documentation.

PetscWrap.create_matrix — Function

create_matrix(
    comm::MPI.Comm = MPI.COMM_WORLD;
    nrows_loc = PETSC_DECIDE,
    ncols_loc = PETSC_DECIDE,
    nrows_glo = PETSC_DECIDE,
    ncols_glo = PETSC_DECIDE,
    autosetup = false,
)

Use autosetup = true to immediatly call set_from_options! and set_up!.

PetscWrap.get_range — Method

get_range(mat::Mat)

Wrapper to MatGetOwnershipRange

However, the result (rstart, rend) is such that mat[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc.MatGetOwnershipRange result where the indexing starts at zero and where rend-1 is last row handled by the local processor.

PetscWrap.get_urange — Method

get_urange(mat::Mat)

Provide a UnitRange from the method get_range.

PetscWrap.mat2file — Function

mat2file(mat::Mat, filename::String, format::PetscViewerFormat = PETSC_VIEWER_ASCII_CSV, type::String = "ascii")

Write a Mat to a file.

PetscWrap.preallocate! — Function

preallocate!(mat::Mat, dnz, onz, warn::Bool = true)

Dispatch preallocation according matrix type (seq or mpiaij for instance). TODO: should use kwargs.

PetscWrap.set_value! — Function

set_value!(mat::Mat, I, J, V, mode = ADD_VALUES)

Set value of mat mat[i, j] = v.

1-based indexing

PetscWrap.set_values! — Function

set_values!(mat::Mat, I, J, V, mode = ADD_VALUES)

Set values of mat in SparseArrays fashion : using COO format: mat[I[k], J[k]] = V[k].

PETSc KSP