Initialization

PetscWrap.PetscFinalize — Method

Wrapper to PetscFinalize

PetscWrap.PetscFinalized — Method

Wrapper to PetscFinalized

Cmd line options: -optionsview - Calls PetscOptionsView() -optionsleft - Prints unused options that remain in the database -objectsdump [all] - Prints list of objects allocated by the user that have not been freed, the option all cause all outstanding objects to be listed -mpidump - Calls PetscMPIDump() -mallocdump - Calls PetscMallocDump(), displays all memory allocated that has not been freed -mallocinfo - Prints total memory usage -mallocview - Prints list of all memory allocated and where

PetscWrap.PetscInitialize — Function

PetscInitialize(cmd_line_args::Bool = true; finalize_atexit = true)

Initialize PETSc.

If cmd_line_args == true, then command line arguments passed to Julia are used as arguments for PETSc (leading to a call to PetscInitializeNoPointers).

Otherwise, if cmd_line_args == false, initialize PETSc without arguments (leading to a call to PetscInitializeNoArguments).

PetscWrap.PetscInitialize — Method

Initialize PETSc with arguments concatenated in a unique string.

PetscWrap.PetscInitialize — Method

Wrapper to `PetscInitializeNoPointers`. Initialize PETCs with arguments.

Implementation

I don't know if I am supposed to use PetscInt or not...

PetscWrap.PetscInitialize — Method

Initialize PETSc with arguments stored in a vector of string

PetscWrap.PetscInitialized — Method

Wrapper to PetscInitialized