PETSc Mat

PetscWrap.PetscMat — Type

A Petsc matrix, actually just a pointer to the actual C matrix

PetscWrap.MatAssemblyBegin — Method

Wrapper to `MatAssemblyBegin`

PetscWrap.MatAssemblyEnd — Method

Wrapper to `MatAssemblyEnd`

PetscWrap.MatCompositeAddMat — Method

MatCompositeAddMat(mat::PetscMat, smat::PetscMat)

Wrapper to MatCompositeAddMat.

PetscWrap.MatCreate — Function

MatCreate(comm::MPI.Comm, mat::PetscMat)

Wrapper to MatCreate

PetscWrap.MatCreateDense — Method

MatCreateDense(comm::MPI.Comm, m::PetscInt, n::PetscInt, M::PetscInt, N::PetscInt)

Wrapper to MatCreateDense. Last argument data is not supported yet (NULL is passed).

PetscWrap.MatCreateVecs — Method

Wrapper to MatCreateVecs

PetscWrap.MatDestroy — Method

MatDestroy(mat::PetscMat)

Wrapper to MatDestroy

PetscWrap.MatGetLocalSize — Method

MatGetLocalSize(mat::PetscMat)

Wrapper to MatGetLocalSize. Return the number of local rows and cols of the matrix (i.e on the processor).

PetscWrap.MatGetOwnershipRange — Method

MatGetOwnershipRange(mat::PetscMat)

Wrapper to MatGetOwnershipRange

The result (rstart, rend) is a Tuple indicating the rows handled by the local processor.

Warning

PETSc indexing starts at zero (so rstart may be zero) and rend-1 is the last row handled by the local processor.

PetscWrap.MatGetOwnershipRangeColumn — Method

MatGetOwnershipRangeColumn(mat::PetscMat)

Wrapper to MatGetOwnershipRangeColumn

The result (cstart, cend) is a Tuple indicating the columns handled by the local processor.

Warning

PETSc indexing starts at zero (so cstart may be zero) and cend-1 is the last column handled by the local processor.

PetscWrap.MatGetSize — Method

MatGetSize(mat::PetscMat)

Wrapper to MatGetSize. Return the number of rows and cols of the matrix (global number).

PetscWrap.MatGetType — Method

MatGetType(mat::PetscMat)

Wrapper to MatGetType. Return the matrix type as a string. See matrix types here: https://petsc.org/release/docs/manualpages/Mat/MatType.html#MatType

PetscWrap.MatMPIAIJSetPreallocation — Method

MatMPIAIJSetPreallocation(mat::PetscMat, dnz::PetscInt, dnnz::Vector{PetscInt}, onz::PetscInt, onnz::Vector{PetscInt})

Wrapper to MatMPIAIJSetPreallocation. dnnz and onnz not tested yet.

Arguments

dnz::PetscInt: number of nonzeros per row in DIAGONAL portion of local submatrix (same value is used for all local rows)

dnnz::Vector{PetscInt}: array containing the number of nonzeros in the various rows of the DIAGONAL portion of the local submatrix (possibly different for each row) or NULL (PETSCNULLINTEGER in Fortran), if d_nz is used to specify the nonzero structure. The size of this array is equal to the number of local rows, i.e 'm'. For matrices that will be factored, you must leave room for (and set) the diagonal entry even if it is zero.

onz::PetscInt: number of nonzeros per row in the OFF-DIAGONAL portion of local submatrix (same value is used for all local rows).

onnz::Vector{PetscInt}: array containing the number of nonzeros in the various rows of the OFF-DIAGONAL portion of the local submatrix (possibly different for each row) or NULL (PETSCNULLINTEGER in Fortran), if o_nz is used to specify the nonzero structure. The size of this array is equal to the number of local rows, i.e 'm'.

PetscWrap.MatMult — Method

MatMult(mat::PetscMat, x::PetscVec, y::PetscVec)

Wrapper to MatMult

PetscWrap.MatSeqAIJSetPreallocation — Method

MatSeqAIJSetPreallocation(mat::PetscMat, nz::PetscInt, nnz::Vector{PetscInt})

Wrapper to MatSeqAIJSetPreallocation.

PetscWrap.MatSetFromOptions — Method

MatSetFromOptions(mat::PetscMat)

Wrapper to MatSetFromOptions

PetscWrap.MatSetOption — Method

Wrapper for MatSetOption

PetscWrap.MatSetSizes — Method

MatSetSizes(mat::PetscMat, nrows_loc::PetscInt, ncols_loc::PetscInt, nrows_glo::PetscInt, ncols_glo::PetscInt)

Wrapper to MatSetSizes

PetscWrap.MatSetType — Method

MatSetType(mat::PetscMat, type::String)

Wrapper for MatSetType. Values for type alors available here: https://petsc.org/release/docs/manualpages/Mat/MatType.html#MatType

PetscWrap.MatSetUp — Method

MatSetUp(mat::PetscMat)

Wrapper to MatSetUp

PetscWrap.MatSetValue — Method

MatSetValue(mat::PetscMat, i::PetscInt, j::PetscInt, v::PetscScalar, mode::InsertMode)

Wrapper to MatSetValue. Indexing starts at 0 (as in PETSc).

Implementation

For an unknow reason, calling PETSc.MatSetValue leads to an "undefined symbol: MatSetValue" error. So this wrapper directly call MatSetValues (anyway, this is what is done in PETSc...)

PetscWrap.MatSetValues — Method

MatSetValues(mat::PetscMat, nI::PetscInt, I::Vector{PetscInt}, nJ::PetscInt, J::Vector{PetscInt}, V::Array{PetscScalar}, mode::InsertMode)

Wrapper to MatSetValues. Indexing starts at 0 (as in PETSc)

PetscWrap.MatView — Function

MatView(mat::PetscMat, viewer::PetscViewer = PetscViewerStdWorld())

Wrapper to MatView