PETSc Viewer

PetscWrap.PetscViewer — Type

PetscViewer(comm::MPI.Comm, filename::String, format::PetscViewerFormat = PETSC_VIEWER_ASCII_CSV, type::String = "ascii", mode::PetscFileMode = FILE_MODE_WRITE)

Constructor for a PetscViewer intended to read/write a matrix or a vector with the supplied type and format.

PetscWrap.PetscViewerASCIIGetStdout — Function

PetscViewerASCIIGetStdout(comm::MPI.Comm = MPI.COMM_WORLD)

Wrapper for PetscViewerASCIIGetStdout

PetscWrap.PetscViewerASCIIOpen — Method

PetscViewerASCIIOpen(comm::MPI.Comm, filename)

Wrapper for PetscViewerASCIIOpen

PetscWrap.PetscViewerCreate — Function

PetscViewerCreate(comm::MPI.Comm = MPI.COMM_WORLD)

Wrapper for PetscViewerCreate

PetscWrap.PetscViewerDestroy — Method

PetscViewerDestroy(viewer::PetscViewer)

Wrapper for PetscViewerDestroy

Warning : from what I understand, all viewers must not be destroyed explicitely using PetscViewerDestroy.

PetscWrap.PetscViewerFileSetMode — Function

PetscViewerFileSetMode(viewer::PetscViewer, mode::PetscFileMode = FILE_MODE_WRITE)

Wrapper for PetscViewerFileSetMode

PetscWrap.PetscViewerFileSetName — Method

PetscViewerFileSetName(viewer::PetscViewer, filename)

Wrapper for PetscViewerFileSetName

PetscWrap.PetscViewerHDF5Open — Method

PetscViewerHDF5Open(comm::MPI.Comm, filename::String, type::PetscFileMode)

Wrapper for PetscViewerHDF5Open

PetscWrap.PetscViewerPopFormat — Method

PetscViewerPopFormat(viewer::PetscViewer)

Wrapper for PetscViewerPopFormat

PetscWrap.PetscViewerPushFormat — Method

PetscViewerPushFormat(viewer::PetscViewer, format::PetscViewerFormat)

Wrapper for PetscViewerPushFormat

PetscWrap.PetscViewerSetType — Method

PetscViewerSetType(viewer::PetscViewer, type::String)

Wrapper for PetscViewerSetType. Values for type alors available here: https://www.mcs.anl.gov/petsc/petsc-current/docs/manualpages/Viewer/PetscViewerType.html#PetscViewerType

PetscWrap.PetscViewerView — Function

PetscViewerView(v::PetscViewer, viewer::PetscViewer = PetscViewerStdWorld())

Wrapper to PetscViewerView