PETSc Vec

Base.setindex! — Method

Base.setindex!(vec::PetscVec, value::Number, row::Integer)

row must be in [1,size(vec)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling VecSetValue fails.

PetscWrap.create_vector — Function

create_vector(nrows, nrows_loc = PETSC_DECIDE)

Create a PetscVec vector of global size (nrows).

PetscWrap.get_range — Method

get_range(vec::PetscVec)

Wrapper to VecGetOwnershipRange

However, the result (rstart, rend) is such that mat[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc.VecGetOwnershipRange result where the indexing starts at zero and where rend-1 is last row handled by the local processor.

PetscWrap.vec2array — Method

vec2array(vec::PetscVec)

Convert a PetscVec into a Julia Array. Allocation is involved in the process since the PetscVec allocated by PETSC is copied into a freshly allocated array. If you prefer not to allocate memory, use VectGetArray and VecRestoreArray

PETSc Mat

Base.setindex! — Method

row and col must be in [1,size(mat)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling MatSetValue fails.

PetscWrap.assemble! — Function

Wrapper to `MatAssemblyBegin` and `MatAssemblyEnd` successively.

PetscWrap.create_matrix — Function

create_matrix(nrows, ncols, nrows_loc = PETSC_DECIDE, ncols_loc = PETSC_DECIDE; auto_setup = false)

Create a PetscMat matrix of global size (nrows, ncols).

Use auto_setup = true to immediatly call set_from_options! and set_up!.

PetscWrap.get_range — Method

get_range(mat::PetscMat)

Wrapper to MatGetOwnershipRange

However, the result (rstart, rend) is such that mat[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc.MatGetOwnershipRange result where the indexing starts at zero and where rend-1 is last row handled by the local processor.

PetscWrap.set_values! — Function

set_values!(mat::PetscMat, I, J, V, mode = ADD_VALUES)

Set values of mat in SparseArrays fashion : using COO format: mat[I[k], J[k]] = V[k].

PETSc KSP