Initialization

PetscWrap.PetscFinalize — Method

Wrapper to PetscFinalize

PetscWrap.PetscInitialize — Function

PetscInitialize(cmd_line_args::Bool = true)

Initialize PETSc.

If cmd_line_args == true, then command line arguments passed to Julia are used as arguments for PETSc (leading to a call to PetscInitializeNoPointers).

Otherwise, if cmd_line_args == false, initialize PETSc without arguments (leading to a call to PetscInitializeNoArguments).

PetscWrap.PetscInitialize — Method

Wrapper to `PetscInitializeNoPointers`. Initialize PETCs with arguments.

Implementation

I don't know if I am supposed to use PetscInt or not...

PetscWrap.PetscInitialize — Method

Initialize PETSc with arguments stored in a vector of string

PetscWrap.PetscInitialize — Method

Initialize PETSc with arguments concatenated in a unique string.