PETSc Mat

PetscWrap.PetscMat — Type

A Petsc matrix, actually just a pointer to the actual C matrix

PetscWrap.MatAssemblyBegin — Method

Wrapper to `MatAssemblyBegin`

PetscWrap.MatAssemblyEnd — Method

Wrapper to `MatAssemblyEnd`

PetscWrap.MatCreate — Method

MatCreate(comm::MPI.Comm, mat::PetscMat)

Wrapper to MatCreate

PetscWrap.MatCreateVecs — Method

Wrapper to MatCreateVecs

PetscWrap.MatDestroy — Method

MatDestroy(mat::PetscMat)

Wrapper to MatDestroy

PetscWrap.MatGetOwnershipRange — Method

MatGetOwnershipRange(mat::PetscMat)

Wrapper to MatGetOwnershipRange

The result (rstart, rend) is a Tuple indicating the rows handled by the local processor.

Warning

PETSc indexing starts at zero (so rstart may be zero) and rend-1 is the last row handled by the local processor.

PetscWrap.MatSetFromOptions — Method

MatSetFromOptions(mat::PetscMat)

Wrapper to MatSetFromOptions

PetscWrap.MatSetSizes — Method

MatSetSizes(mat::PetscMat, nrows_loc, ncols_loc, nrows_glo, ncols_glo)

Wrapper to MatSetSizes

PetscWrap.MatSetUp — Method

MatSetUp(mat::PetscMat)

Wrapper to MatSetUp

PetscWrap.MatSetValue — Method

MatSetValue(mat::PetscMat, i::PetscInt, j::PetscInt, v::PetscScalar, mode::InsertMode)

Wrapper to MatSetValue. Indexing starts at 0 (as in PETSc).

Implementation

For an unknow reason, calling PETSc.MatSetValue leads to an "undefined symbol: MatSetValue" error. So this wrapper directly call MatSetValues (anyway, this is what is done in PETSc...)

PetscWrap.MatSetValues — Method

MatSetValues(mat::PetscMat, nI::PetscInt, I::Vector{PetscInt}, nJ::PetscInt, J::Vector{PetscInt}, V::Array{PetscScalar}, mode::InsertMode)

Wrapper to MatSetValues. Indexing starts at 0 (as in PETSc)

PetscWrap.MatView — Function

MatView(mat::PetscMat, viewer::PetscViewer = PetscViewerStdWorld())

Wrapper to MatView