PETSc Mat

PetscWrap.PetscMat — Type

A Petsc matrix, actually just a pointer to the actual C matrix

Base.setindex! — Method

row and col must be in [1,size(mat)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling MatSetValue fails.

PetscWrap.MatAssemblyBegin — Method

Wrapper to MatAssemblyBegin

PetscWrap.MatAssemblyEnd — Method

Wrapper to MatAssemblyEnd

PetscWrap.MatCreate — Method

MatCreate(comm, mat::PetscMat)

Wrapper to MatCreate

PetscWrap.MatCreateVecs — Method

Wrapprt to MatCreateVecs

PetscWrap.MatDestroy — Method

MatDestroy(mat::PetscMat)

Wrapper to MatDestroy

PetscWrap.MatGetOwnershipRange — Method

MatGetOwnershipRange(mat::PetscMat)

Wrapper to MatGetOwnershipRange

However, the result (rstart, rend) is such that mat[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc result where the indexing starts at one and where rend-1 is last row handled by the local processor.

PetscWrap.MatSetFromOptions — Method

MatSetFromOptions(mat::PetscMat)

Wrapper to MatSetFromOptions

PetscWrap.MatSetSizes — Method

MatSetSizes(mat::PetscMat, nrows_loc, ncols_loc, nrows_glo, ncols_glo)

Wrapper to MatSetSizes

PetscWrap.MatSetUp — Method

MatSetUp(mat::PetscMat)

Wrapper to MatSetUp

PetscWrap.MatSetValues — Method

MatSetValues(mat::PetscMat, I::Vector{PetscInt}, J::Vector{PetscInt}, V::Array{PetscScalar}, mode::InsertMode)

Wrapper to MatSetValues. Indexing starts at 1 (Julia)

PetscWrap.MatView — Function

MatView(mat::PetscMat, viewer::PetscViewer = C_NULL)

Wrapper to MatView

PetscWrap.create_matrix — Function

create_matrix(nrows, ncols, nrows_loc = PETSC_DECIDE, ncols_loc = PETSC_DECIDE)

Create a PetscMat matrix of global size (nrows, ncols).