PETSc Vec

PetscWrap.VecAssemblyBegin — Method

VecAssemblyBegin(vec::PetscVec)

Wrapper to VecAssemblyBegin

PetscWrap.VecAssemblyEnd — Method

VecAssemblyEnd(vec::PetscVec)

Wrapper to VecAssemblyEnd

PetscWrap.VecCreate — Method

VecCreate(comm::MPI.Comm, vec::PetscVec)

Wrapper to VecCreate

PetscWrap.VecDestroy — Method

VecDestroy(vec::PetscVec)

Wrapper to VecDestroy

PetscWrap.VecDuplicate — Method

VecDuplicate(vec::PetscVec)

Wrapper for VecDuplicate, except that it returns the new vector instead of taking it as an input.

PetscWrap.VecGetArray — Function

VecGetArray(vec::PetscVec, own = false)

Wrapper for VecGetArray.

Warning

I am not confortable at all with memory management, both on the C side and on the Julia side. Use this at you own risk.

According to Julia documentation, own optionally specifies whether Julia should take ownership of the memory, calling free on the pointer when the array is no longer referenced."

PetscWrap.VecGetLocalSize — Method

VecGetLocalSize(vec::PetscVec)

Wrapper for VecGetLocalSize

PetscWrap.VecGetOwnershipRange — Method

VecGetOwnershipRange(vec::PetscVec)

Wrapper to VecGetOwnershipRange

The result (rstart, rend) is a Tuple indicating the rows handled by the local processor.

Warning

PETSc indexing starts at zero (so rstart may be zero) and rend-1 is the last row handled by the local processor.

PetscWrap.VecGetSize — Method

VecGetSize(vec::PetscVec)

Wrapper for VecGetSize

PetscWrap.VecRestoreArray — Method

VecRestoreArray(vec::PetscVec, array_ref)

Wrapper for VecRestoreArray. array_ref is obtained from VecGetArray

PetscWrap.VecSetFromOptions — Method

VecSetFromOptions(vec::PetscVec)

Wrapper to VecSetFromOptions

PetscWrap.VecSetSizes — Method

VecSetSizes(vec::PetscVec, nrows_loc, nrows_glo)

Wrapper to VecSetSizes

PetscWrap.VecSetUp — Method

VecSetUp(vec::PetscVec)

Wrapper to VecSetUp

PetscWrap.VecSetValue — Function

VecSetValue(vec::PetscVec, i::PetscInt, v::PetscScalar, mode::InsertMode = INSERT_VALUES)

Wrapper to VecSetValue. Indexing starts at 0 (as in PETSc).

Implementation

For an unknow reason, calling PETSc.VecSetValue leads to an "undefined symbol: VecSetValue" error. So this wrapper directly call VecSetValues (anyway, this is what is done in PETSc...)

PetscWrap.VecSetValues — Function

VecSetValues(vec::PetscVec, nI::PetscInt, I::Vector{PetscInt}, V::Array{PetscScalar}, mode::InsertMode = INSERT_VALUES)

Wrapper to VecSetValues. Indexing starts at 0 (as in PETSc)

PetscWrap.VecView — Method

VecView(vec::PetscVec, viewer::PetscViewer)

Wrapper to VecView. Currently leads to a fatal error when using with viewer = PetscViewerStdWorld(). Use VecView(vec) instead.

PetscWrap.VecView — Method

VecView(vec::PetscVec)

Wrapper to VecView with the default PETSc viewer. This a fallback for the currently buggy VecView(vec::PetscVec, viewer::PetscViewer)