PETSc Vec

Base.setindex! — Method

Base.setindex!(vec::PetscVec, value::Number, row::Integer)

row must be in [1,size(vec)], i.e indexing starts at 1 (Julia).

Implementation

For some unkwnown reason, calling VecSetValue fails.

PetscWrap.VecAssemblyBegin — Method

VecAssemblyBegin(vec::PetscVec)

Wrapper to VecAssemblyBegin

PetscWrap.VecAssemblyEnd — Method

VecAssemblyEnd(vec::PetscVec)

Wrapper to VecAssemblyEnd

PetscWrap.VecCreate — Method

VecCreate(comm, vec::PetscVec)

Wrapper to VecCreate

PetscWrap.VecDestroy — Method

VecDestroy(vec::PetscVec)

Wrapper to VecDestroy

PetscWrap.VecDuplicate — Method

VecDuplicate(vec::PetscVec)

Wrapper for VecDuplicate, except that it returns the new vector instead of taking it as an input.

PetscWrap.VecGetArray — Function

VecGetArray(vec::PetscVec, own = false)

Wrapper for VecGetArray.

Warning

I am not confortable at all with memory management, both on the C side and on the Julia side. Use this at you own risk.

According to Julia documentation, own optionally specifies whether Julia should take ownership of the memory, calling free on the pointer when the array is no longer referenced."

PetscWrap.VecGetLocalSize — Method

VecGetLocalSize(vec::PetscVec)

Wrapper for VecGetLocalSize

PetscWrap.VecGetOwnershipRange — Method

VecGetOwnershipRange(vec::PetscVec)

Wrapper to VecGetOwnershipRange

However, the result (rstart, rend) is such that mat[rstart:rend] are the rows handled by the local processor. This is different from the default PETSc result where the indexing starts at one and where rend-1 is last row handled by the local processor.

PetscWrap.VecGetSize — Method

VecGetSize(vec::PetscVec)

Wrapper for VecGetSize

PetscWrap.VecRestoreArray — Method

VecRestoreArray(vec::PetscVec, array_ref)

Wrapper for VecRestoreArray. array_ref is obtained from VecGetArray

PetscWrap.VecSetFromOptions — Method

VecSetFromOptions(vec::PetscVec)

Wrapper to VecSetFromOptions

PetscWrap.VecSetSizes — Method

VecSetSizes(vec::PetscVec, nrows_loc, nrows_glo)

Wrapper to VecSetSizes

PetscWrap.VecSetUp — Method

VecSetUp(vec::PetscVec)

Wrapper to VecSetUp

PetscWrap.VecSetValues — Function

VecSetValues(vec::PetscVec, I::Vector{PetscInt}, V::Array{PetscScalar}, mode::InsertMode = INSERT_VALUES)

Wrapper to VecSetValues. Indexing starts at 1 (Julia)

PetscWrap.VecView — Function

VecView(vec::PetscVec, viewer::PetscViewer = C_NULL)

Wrapper to VecView

PetscWrap.create_vector — Function

create_vector(nrows, nrows_loc = PETSC_DECIDE)

Create a PetscVec matrix of global size (nrows).

PetscWrap.vec2array — Method

vec2array(vec::PetscVec)

Convert a PetscVec into a Julia Array. Allocation is involved in the process since the PetscVec allocated by PETSC is copied into a freshly allocated array. If you prefer not to allocate memory, use VectGetArray and VecRestoreArray